Move sustainability_labels_to_certificates to utils
